# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: CommercePulse analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: platform
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/fulfill:
    post:
      operationId: fulfillOrder
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/send-receipt:
    post:
      operationId: sendOrderReceipt
      tags: [orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Receipt send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendReceiptResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Products with performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductPerformance"
    post:
      operationId: createProduct
      tags: [products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPerformance"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPerformance"
    delete:
      operationId: deleteProduct
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RangeParam:
      in: query
      name: range
      schema:
        type: string
        enum: [7d, 14d, 30d, 90d]
        default: 30d
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    OrderRow:
      type: object
      required:
        [
          id,
          orderNumber,
          platform,
          status,
          totalAmount,
          profit,
          orderedAt,
          productSummary,
          itemCount,
        ]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        platform: { type: string }
        status: { type: string }
        totalAmount: { type: number }
        profit: { type: number }
        orderedAt: { type: string }
        productSummary: { type: string }
        itemCount: { type: integer }
    OrderListResponse:
      type: object
      required: [orders, summary]
      properties:
        summary:
          type: object
          required: [revenue, profit, count]
          properties:
            revenue: { type: number }
            profit: { type: number }
            count: { type: integer }
        orders:
          type: array
          items: { $ref: "#/components/schemas/OrderRow" }
    OrderItemDetail:
      type: object
      required:
        [id, name, quantity, unitPrice, unitCost, lineTotal]
      properties:
        id: { type: string }
        productId: { type: ["string", "null"] }
        name: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        unitCost: { type: number }
        lineTotal: { type: number }
    OrderSummary:
      type: object
      required:
        [
          id,
          orderNumber,
          platform,
          status,
          totalAmount,
          orderedAt,
        ]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        platform: { type: string }
        status: { type: string }
        totalAmount: { type: number }
        orderedAt: { type: string }
    SendReceiptResponse:
      type: object
      required: [sent, orderId, recipientEmail]
      properties:
        sent: { type: boolean }
        orderId: { type: string }
        recipientEmail: { type: string }
        reason: { type: string }
    OrderDetail:
      type: object
      required: [order, items]
      properties:
        order:
          type: object
          required:
            [
              id,
              orderNumber,
              platform,
              status,
              totalAmount,
              subtotal,
              shipping,
              tax,
              profit,
              orderedAt,
            ]
          properties:
            id: { type: string }
            orderNumber: { type: string }
            platform: { type: string }
            status: { type: string }
            totalAmount: { type: number }
            subtotal: { type: number }
            shipping: { type: number }
            tax: { type: number }
            profit: { type: number }
            orderedAt: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemDetail" }
        customer:
          type: ["object", "null"]
          properties:
            id: { type: ["string", "null"] }
            name: { type: string }
            email: { type: string }
            phone: { type: ["string", "null"] }
    ProductPerformance:
      type: object
      required:
        [
          id,
          name,
          category,
          price,
          cogs,
          stock,
          status,
          lowStock,
          unitsSold,
          revenue,
          profit,
          margin,
          roas,
        ]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        cogs: { type: number }
        stock: { type: integer }
        status: { type: string }
        lowStock: { type: boolean }
        unitsSold: { type: integer }
        revenue: { type: number }
        profit: { type: number }
        margin: { type: number }
        roas: { type: number }
    CreateProductBody:
      type: object
      required: [name, price]
      properties:
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        price: { type: number }
        cogs: { type: number }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
    UpdateProductBody:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        price: { type: number }
        cogs: { type: number }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
        status: { type: string }
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }
...